+++ Lab 12/Project Folder/ADC.h
@@ -1,9 +1,12 @@
 #ifndef __ADC_H_
 #define __ADC_H_
 
+#include "stm32f446xx.h"
+
 #define VREF 3.3
 
-int Read_ADC(void);
+
+uint8_t Read_ADC(void);
 void adc_init(void);
 
 #endif
